+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## [8.0.3] - 2021-02-20
 
+### Added
+
+- Added pt_PT language
+
 ### Changed
 
 - Updated composer