+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - Update to Laravel commit: f0de9fd9967d4e1b4427d8458bf8983bc2cde201
 - Upgrade to Laravel Mix 6
 - Fixed 2fa/admin issue (https://github.com/rappasoft/laravel-boilerplate/pull/1488)
+- Update pt_BR language
 
 ### Removed