+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 - Added pt_PT language
 - Added RO language
+- Added missing cors middleware
 
 ### Changed